requirements.txt
javalang==0.13.0
jellyfish==1.0.1
numpy==1.24.4
pandas==2.0.3
PyYAML==6.0.1
pyfiglet==1.0.2
torch==2.2.1
tqdm==4.66.1
wordninja==2.0.0
xmltodict==0.13.0
